src/main/java/org/example/angularfejavabe/service/UserService.java
package org.example.angularfejavabe.service;

import org.example.angularfejavabe.model.User;
import org.example.angularfejavabe.repo.UserRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class UserService {
    
    
    private final UserRepository userRepository;
    
    public UserService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }
    
    public List<User> findAll() {
        return userRepository.findAll();
    }
    
    public User save(User user) {
        return userRepository.save(user);
    }
}
